Тянет все таски за месяц  и отрабатывает в ChoosedDay таски с облака
