+++ src/components/ChoosedDay/ChoosedDay.jsx
@@ -1,11 +1,10 @@
-import React, { useEffect } from 'react';
+import React, { useEffect, useState } from 'react';
 import { TasksColumnsList } from './TasksColumnsList/TasksColumnsList';
 import { TasksColumnsListWrapper } from './ChoosedDay.styled';
 import { DayCalendarHead } from './DayCalendarHead/DayCalendarHead';
 import { useDispatch, useSelector } from 'react-redux';
 
 import { selectArrTasks } from 'redux/tasks/tasks-selectors';
-import { fetchTasks } from 'redux/tasks/tasks-operations';
 
 import {
   selectAddTaskOpen,
@@ -20,6 +19,9 @@ import {
 } from 'redux/modal/globalSlice';
 
 const ChoosedDay = () => {
+
+  const [tasksFilter, setTasksFilter] = useState([]);
+
   const tasksMonth = useSelector(selectArrTasks);
   const modalAddState = useSelector(selectAddTaskOpen);
   const modalEditState = useSelector(selectUpDateTaskModal);
@@ -27,17 +29,8 @@ const ChoosedDay = () => {
 
   const dispatch = useDispatch();
 
-  useEffect(() => {
-    const currentDate = new Date();
-    const year = String(currentDate.getFullYear());
-    const month = String(currentDate.getMonth() + 1).padStart(2, '0');
 
-    dispatch(fetchTasks({ month, year }));
 
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
- 
   const closeModal = () => {
     dispatch(closeModalAddTask());
   };
@@ -48,15 +41,27 @@ const ChoosedDay = () => {
   const deleteTask = () => {
     closeDeleteModal();
   };
- 
-  const chooseDay = (date) => {
-    console.log(date);
-  };
+
+  const chooseDay = date => {
+    const { day, month, year } = date;
+  
+      const filteredTasks = tasksMonth.filter(({start, end}) => {
+      
+        const chosenDate = `${year}/${month}/${day}`;
+
+        return start <= chosenDate && end >= chosenDate;
+      });
+    
+      setTasksFilter(filteredTasks);
+      
+    };
+    console.log('ChoosenDay tasks', tasksFilter);
+  
 
   return (
     <TasksColumnsListWrapper>
       <DayCalendarHead clickChooseDay={chooseDay} />
-      <TasksColumnsList tasks={tasksMonth} />
+      <TasksColumnsList tasks={tasksFilter} />
       {modalAddState && <Modal closeModal={closeModal} typeOfModal={'add'} />}
       {modalEditState && (
         <Modal closeModal={closeEditModal} typeOfModal={'edit'} />
